Separate logic from Markup
